pyproject.toml
[tool.poetry]
name = "django-training"
version = "0.1.0"
description = "bld.ai backend internship projects"
authors = ["AmmarDab3an <ammar.dabaan99@gmail.com>"]
readme = "README.md"
packages = [{include = "django_training"}]

[tool.poetry.dependencies]
python = "^3.10"
Django = "^4.1.2"
django-model-utils = "^4.2.0"
django-debug-toolbar = "^3.7.0"
django-imagekit = "^4.1.0"
Pillow = "^9.2.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
